refactor: #392
도메인 서비스 로직과 어플리케이션 서비스 로직 분리
