+++ ticketing/src/main/java/com/goti/ticketing/game/service/domain/GameTicketingStatusService.java
@@ -6,4 +6,6 @@ import com.goti.ticketing.domain.entity.game.GameTicketingStatusEntity;
 public interface GameTicketingStatusService {
 
 	GameTicketingStatusEntity create(GameScheduleEntity gameSchedule);
+
+	GameTicketingStatusEntity getLocked(GameScheduleEntity gameSchedule);
 }
