+++ ticketing/src/main/java/com/goti/ticketing/order/service/application/OrderPaymentConfirmService.java
@@ -17,7 +17,7 @@ import com.goti.ticketing.domain.entity.order.OrderItemEntity;
 import com.goti.ticketing.domain.entity.seat.SeatHoldEntity;
 import com.goti.exception.CustomException;
 import com.goti.global.validation.Preconditions;
-import com.goti.ticketing.game.service.domain.TicketingStatusUpdateService;
+import com.goti.ticketing.game.service.application.TicketingStatusUpdateService;
 import com.goti.ticketing.order.dto.response.OrderPaymentConfirmResponse;
 import com.goti.ticketing.order.repository.OrderItemRepository;
 import com.goti.ticketing.order.repository.OrderRepository;
